+++ ci/deploy.sh
@@ -1,5 +1,5 @@
 #!/usr/bin/env bash
 
-if [[ $TRAVIS_BRANCH =~ .*master.* ]]; then
-  ./gradlew deployToBintray
+if [[ ${TRAVIS_BRANCH} =~ .*master.* ]]; then
+  ./gradlew deployToBintray deployToGitHub
 fi