+++ library-service/src/androidTest/java/universum/studios/android/officium/service/ServiceApiProviderTest.java
@@ -17,12 +17,18 @@
  * =================================================================================================
  */
 package universum.studios.android.officium.service; 
+import android.support.annotation.NonNull;
 import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import universum.studios.android.test.BaseInstrumentedTest;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.core.IsNot.not;
+import static org.hamcrest.core.IsNull.notNullValue;
 
 /**
  * @author Martin Albedinsky
@@ -34,7 +40,37 @@ public final class ServiceApiProviderTest extends BaseInstrumentedTest {
 	private static final String TAG = "ServiceApiProviderTest";
 
     @Test
-	public void test() {
-		// todo:: implement test
+	public void testGetApi() {
+		final ServiceApiProvider<TestApiImpl> provider = new TestProvider();
+	    assertThat(provider.getApi(), is(notNullValue()));
+	    assertThat(provider.getApi(), is(provider.getApi()));
+	}
+
+    @Test
+	public void testInvalidateApi() {
+		final ServiceApiProvider<TestApiImpl> provider = new TestProvider();
+	    final TestApi api = provider.getApi();
+	    provider.invalidateApi();
+	    assertThat(provider.getApi(), is(not(api)));
+	}
+
+    @Test
+	public void testInvalidateApiNotInitialized() {
+	    new TestProvider().invalidateApi();
+	}
+
+	private interface TestApi {
+	}
+
+	private final class TestApiImpl implements TestApi {
+	}
+
+	private final class TestProvider extends ServiceApiProvider<TestApiImpl> {
+
+		@NonNull
+		@Override
+		protected TestApiImpl onCreateApi() {
+			return new TestApiImpl();
+		}
 	}
 }