+++ library-sync/src/androidTest/java/universum/studios/android/officium/sync/SyncHandlerTest.java
@@ -17,12 +17,21 @@
  * =================================================================================================
  */
 package universum.studios.android.officium.sync; 
+import android.accounts.Account;
+import android.content.Context;
+import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.support.test.runner.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
 import universum.studios.android.test.BaseInstrumentedTest;
+
+import static org.hamcrest.MatcherAssert.assertThat;
+import static org.hamcrest.core.Is.is;
+import static org.hamcrest.core.IsNull.notNullValue;
+import static org.hamcrest.core.IsNull.nullValue;
 
 /**
  * @author Martin Albedinsky
@@ -34,7 +43,122 @@ public final class SyncHandlerTest extends BaseInstrumentedTest {
 	private static final String TAG = "SyncHandlerTest";
 
     @Test
-	public void test() {
-		// todo:: implement test
+	public void testInstantiationWithId() {
+		assertThat(new TestHandler(1).getTaskId(), is(1));
+    }
+
+    @Test
+	public void testInstantiationWithIdAndRequestClass() {
+	    assertThat(new TestHandler(1, TestRequest.class).getTaskId(), is(1));
+    }
+
+	@Test
+	public void testHandleSync() {
+		final TestHandler handler = new TestHandler(1, TestRequest.class);
+		handler.handleSync(mContext, new SyncOperation.Builder()
+				.account(new Account("TestAccount", "TestType"))
+				.authority("test.authority")
+				.task(new SyncTask.Builder<TestRequest>(1).request(new TestRequest(12)).build())
+				.build()
+		);
+		assertThat(handler.hasBeenOnHandleSyncInvoked(), is(true));
+		assertThat(handler.hasBeenOnSyncErrorInvoked(), is(false));
+		final TestRequest request = handler.getOnHandleSyncRequest();
+		assertThat(request, is(notNullValue()));
+		assertThat(request.count, is(12));
+	}
+
+	@Test
+	public void testHandleSyncWithoutRequest() {
+		final TestHandler handler = new TestHandler(1);
+		handler.handleSync(mContext, new SyncOperation.Builder()
+				.account(new Account("TestAccount", "TestType"))
+				.authority("test.authority")
+				.task(SyncTask.EMPTY)
+				.build()
+		);
+		assertThat(handler.hasBeenOnHandleSyncInvoked(), is(true));
+		assertThat(handler.hasBeenOnSyncErrorInvoked(), is(false));
+		assertThat(handler.getOnHandleSyncRequest(), is(nullValue()));
+	}
+
+	@Test
+	public void testHandleRequestWithOccurredError() {
+		final TestHandler handler = new TestHandler(1, TestRequest.class);
+		handler.setExceptionToThrow(new IllegalStateException());
+		handler.handleSync(mContext, new SyncOperation.Builder()
+				.account(new Account("TestAccount", "TestType"))
+				.authority("test.authority")
+				.task(SyncTask.EMPTY)
+				.build()
+		);
+		assertThat(handler.hasBeenOnHandleSyncInvoked(), is(true));
+		assertThat(handler.hasBeenOnSyncErrorInvoked(), is(true));
+		assertThat(handler.isOnSyncErrorEqualExceptionToThrow(), is(true));
+	}
+
+    private static final class TestRequest implements SyncTask.Request {
+
+	    int count;
+
+	    TestRequest() {
+	    }
+
+	    TestRequest(int count) {
+		    this.count = count;
+	    }
+    }
+
+	private static final class TestHandler extends SyncHandler<TestRequest, Void> {
+
+		private TestRequest onHandleSyncRequest;
+		private boolean onHandleSyncInvoked;
+		private Exception exceptionToThrow;
+		private Exception onSyncError;
+
+		TestHandler(int taskId) {
+			super(taskId);
+		}
+
+		TestHandler(int taskId, @Nullable Class<TestRequest> classOfRequest) {
+			super(taskId, classOfRequest);
+		}
+
+		void setExceptionToThrow(Exception toThrow) {
+			this.exceptionToThrow = toThrow;
+		}
+
+		@Nullable
+		@Override
+		protected Void onHandleSync(@NonNull Context context, @NonNull SyncOperation syncOperation, @Nullable TestRequest syncRequest) throws Exception {
+			this.onHandleSyncInvoked = true;
+			this.onHandleSyncRequest = syncRequest;
+			if (exceptionToThrow != null) {
+				throw exceptionToThrow;
+			}
+			return null;
+		}
+
+		boolean hasBeenOnHandleSyncInvoked() {
+			return onHandleSyncInvoked;
+		}
+
+		TestRequest getOnHandleSyncRequest() {
+			return onHandleSyncRequest;
+		}
+
+		@Override
+		protected void onSyncError(@NonNull Context context, @NonNull SyncOperation syncOperation, @Nullable TestRequest syncRequest, @NonNull Exception error) {
+			super.onSyncError(context, syncOperation, syncRequest, error);
+			this.onSyncError = error;
+		}
+
+		boolean hasBeenOnSyncErrorInvoked() {
+			return onSyncError != null;
+		}
+
+		boolean isOnSyncErrorEqualExceptionToThrow() {
+			return onSyncError == exceptionToThrow;
+		}
 	}
 }